tf/aws/requirements.txt
tensorflow
numpy
Pillow
requests
matplotlib
